Add files via upload
The aim of the code in this folder, "Product Landing Page", is to recreate the contents of this image.
Due to learning reasons, this cloned page was intentionally made to be non-functional. For instance, the non-clickable links etc.
The page was made with the sole purpose of just recreating the design, the goal being to test/improve my CSS skills.